requirements.txt
pandas
datasets
llama-cpp-python
ollama
torch
transformers
tqdm
git+https://github.com/unslothai/unsloth.git#egg=unsloth[colab-new] # this is the nighly version
